feat : 기존 생성자 가시성 제한
