+++ screennameviewer/compose/src/main/java/com/donglab/screennameviewer/publicapi/config/ScreenNameOverlayConfig.kt
@@ -6,7 +6,7 @@ import android.view.Gravity
 /**
  * ScreenNameViewer의 UI 설정을 담당하는 설정 클래스
  */
-data class ScreenNameOverlayConfig(
+data class ScreenNameOverlayConfig internal constructor(
     val textSize: Float,
     val textColor: Int,
     val backgroundColor: Int,
