+++ screennameviewer/compose/src/main/java/com/donglab/screennameviewer/publicapi/extensions/ScreenNameViewerExtensions.kt
@@ -9,7 +9,7 @@ import androidx.navigation.NavController
 import com.donglab.screennameviewer.internal.compose.tracker.ComposeScreenNameTracker
 import com.donglab.screennameviewer.internal.util.findActivityContext
 import com.donglab.screennameviewer.internal.compose.viewer.ComposeRouteViewerImpl
-import com.donglab.screennameviewer.publicapi.ScreenNameViewer
+import com.donglab.screennameviewer.publicapi.viewer.ScreenNameViewer
 
 /**
  * ComposeScreenNameTracker를 활성화하는 Composable 함수입니다.
